LED sortof works now. many bugs remain in ECsevensegment etc...
git-svn-id: https://ktechlab.svn.sourceforge.net/svnroot/ktechlab/trunk@339 76876f85-f02d-0410-881d-fd707285c4ba

+++ src/electronics/components/ecsevensegment.h
@@ -35,8 +35,8 @@ private:
 	virtual void drawShape( QPainter &p );
 	
 	bool m_bCommonCathode;
-	double lastUpdatePeriod;
-	double avg_brightness[8];
+	uint lastUpdatePeriod;
+	uint avg_brightness[8];
 	uint last_brightness[8];
 	Diode *m_diodes[8];
 	ECNode *m_nodes[8];
